requirements.txt
fastapi
uvicorn
streamlit
snowflake
pydantic
snowflake-connector-python
passlib[bcrypt]
python-dotenv
requests
fpdf
pydantic
logging
pinecone-client
openai==0.28
google-api-python-client
youtube-transcript-api
nltk
